chore: refactor RoleManagementAccounts
