+++ helpers/role-management/src/processor.rs
@@ -5,7 +5,7 @@ use solana_program::bpf_loader_upgradeable::UpgradeableLoaderState;
 use solana_program::entrypoint::ProgramResult;
 use solana_program::program_error::ProgramError;
 use solana_program::pubkey::Pubkey;
-use solana_program::{bpf_loader_upgradeable, msg};
+use solana_program::{bpf_loader_upgradeable, msg, system_program};
 
 use crate::instructions::RoleManagementInstructionInputs;
 use crate::seed_prefixes;
@@ -493,9 +493,13 @@ impl<'a> TryFrom<&'a [AccountInfo<'a>]> for RoleManagementAccounts<'a> {
 
     fn try_from(value: &'a [AccountInfo<'a>]) -> Result<Self, Self::Error> {
         let account_iter = &mut value.iter();
+        let system_account = next_account_info(account_iter)?;
+        if !system_program::check_id(system_account.key) {
+            return Err(ProgramError::IncorrectProgramId);
+        }
 
         Ok(Self {
-            system_account: next_account_info(account_iter)?,
+            system_account,
             payer: next_account_info(account_iter)?,
             payer_roles_account: next_account_info(account_iter)?,
             resource: next_account_info(account_iter)?,
@@ -774,5 +778,59 @@ mod tests {
                 "Invalid result for bit pattern {bits}",
             );
         }
+    }
+
+    #[test]
+    fn test_accounts_from_role_management_accounts() {
+        let key = Pubkey::new_unique();
+
+        let mut system_account_lamports = 1;
+        let mut system_account_data = [1, 2, 3];
+        let system_account = AccountInfo::new(
+            &system_program::ID,
+            true,
+            true,
+            &mut system_account_lamports,
+            &mut system_account_data,
+            &key,
+            true,
+            1,
+        );
+
+        let mut lamports = 2;
+        let mut data = [1, 2, 3];
+        let dummy_account =
+            AccountInfo::new(&key, true, true, &mut lamports, &mut data, &key, true, 1);
+
+        let accounts = [
+            system_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+        ];
+
+        let res = RoleManagementAccounts::try_from(accounts.as_slice());
+        assert!(res.is_ok());
+
+        // change system account to fail
+        let accounts = [
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account.clone(),
+            dummy_account,
+        ];
+        let res = RoleManagementAccounts::try_from(accounts.as_slice());
+        assert!(res.is_err());
+        assert_eq!(res.unwrap_err(), ProgramError::IncorrectProgramId);
     }
 }
